+++ app/controllers/integral/application_controller.rb
@@ -17,6 +17,13 @@ module Integral
       super
     end
 
+    helper_method :category_path
+
+    # @return [String] path to category page
+    def category_path(category)
+      "/#{Integral.blog_namespace}/#{category.slug}"
+    end
+
     private
 
     # Actions that should be carried out before rendering