+++ app/controllers/integral/backend/base_controller.rb
@@ -29,6 +29,29 @@ module Integral
       # GET /
       # Lists all resources
       def index
+        # TODO: This default behaviour will switch to 'list' action
+
+        respond_to do |format|
+          format.html do
+            set_grid
+          end
+
+          format.json do
+            if params[:gridview].present?
+              set_grid
+              render json: { content: render_to_string(partial: "integral/backend/#{controller_name}/grid", locals: { grid: @grid }) }
+            else
+              respond_to_record_selector
+            end
+          end
+        end
+      end
+
+      # GET /list
+      # Lists all resources
+      def list
+        add_breadcrumb I18n.t('integral.navigation.list'), "new_backend_#{controller_name.singularize}_path".to_sym
+
         respond_to do |format|
           format.html do
             set_grid
@@ -86,12 +109,21 @@ module Integral
       # DELETE /:id
       def destroy
         if @resource.destroy
-          respond_successfully(notification_message('delete_success'), send("backend_#{controller_name}_path"))
+          respond_to do |format|
+            format.html { respond_successfully(notification_message('delete_success'), send("backend_#{controller_name}_path")) }
+            format.js { head :no_content }
+          end
         else
-          error_message = @resource.errors.full_messages.to_sentence
-          flash[:error] = "#{notification_message('delete_failure')} - #{error_message}"
+          respond_to do |format|
+            format.html do
+              error_message = @resource.errors.full_messages.to_sentence
+              flash[:error] = "#{notification_message('delete_failure')} - #{error_message}"
 
-          redirect_to send("backend_#{controller_name}_path")
+              redirect_to send("backend_#{controller_name}_path")
+            end
+            format.js { head :unprocessable_entity }
+          end
+
         end
       end