+++ db/migrate/20191203090008_add_image_to_integral_categories.rb
@@ -1,0 +1,6 @@
+class AddImageToIntegralCategories < ActiveRecord::Migration[5.2]
+  def change
+    add_column :integral_categories, :image_id, :bigint
+    add_index :integral_categories, :image_id
+  end
+end